+++ packages/crdt_lf/lib/crdt_lf.dart
@@ -13,6 +13,7 @@ export 'src/dag/node.dart';
 export 'src/document.dart';
 export 'src/exceptions/exceptions.dart';
 export 'src/frontiers/frontiers.dart';
+export 'src/handler/fugue_list/handler.dart';
 export 'src/handler/fugue_text/handler.dart';
 export 'src/handler/handler.dart';
 export 'src/handler/list/handler.dart';